use bigint for unique ID

+++ create_Extractor_HTML_tables.sql
@@ -2,11 +2,14 @@
 -- the SEC database which is extracted from the files
 -- downloaded from the SEC FTP server.
 
+-- see link below for hints
+-- https://wiki.postgresql.org/wiki/Don%27t_Do_This
+
 DROP TABLE IF EXISTS html_extracts.sec_filing_id CASCADE ;
 
 CREATE TABLE html_extracts.sec_filing_id
 (
-    filing_ID integer GENERATED ALWAYS AS IDENTITY UNIQUE,
+    filing_ID bigint GENERATED ALWAYS AS IDENTITY UNIQUE,
 	cik TEXT NOT NULL,
 	company_name TEXT NOT NULL,
 	file_name TEXT NOT NULL,
@@ -32,8 +35,8 @@ DROP TABLE IF EXISTS html_extracts.sec_cash_flows_data ;
 
 CREATE TABLE html_extracts.sec_bal_sheet_data
 (
-    filing_data_ID integer GENERATED ALWAYS AS IDENTITY UNIQUE,
-	filing_ID integer REFERENCES html_extracts.sec_filing_id (filing_ID) ON DELETE CASCADE,
+    filing_data_ID bigint GENERATED ALWAYS AS IDENTITY UNIQUE,
+	filing_ID bigint REFERENCES html_extracts.sec_filing_id (filing_ID) ON DELETE CASCADE,
 	html_label TEXT NOT NULL,
     html_value TEXT NOT NULL,
 	/* period_begin DATE NOT NULL, */
@@ -59,8 +62,8 @@ CREATE INDEX idx_bal_sheet ON html_extracts.sec_bal_sheet_data USING GIN (tsv);
 
 CREATE TABLE html_extracts.sec_stmt_of_ops_data
 (
-    filing_data_ID integer GENERATED ALWAYS AS IDENTITY UNIQUE,
-	filing_ID integer REFERENCES html_extracts.sec_filing_id (filing_ID) ON DELETE CASCADE,
+    filing_data_ID bigint GENERATED ALWAYS AS IDENTITY UNIQUE,
+	filing_ID bigint REFERENCES html_extracts.sec_filing_id (filing_ID) ON DELETE CASCADE,
 	html_label TEXT NOT NULL,
     html_value TEXT NOT NULL,
 	/* period_begin DATE NOT NULL, */
@@ -86,8 +89,8 @@ CREATE INDEX idx_stmt_of_ops ON html_extracts.sec_stmt_of_ops_data USING GIN (ts
 
 CREATE TABLE html_extracts.sec_cash_flows_data
 (
-    filing_data_ID integer GENERATED ALWAYS AS IDENTITY UNIQUE,
-	filing_ID integer REFERENCES html_extracts.sec_filing_id (filing_ID) ON DELETE CASCADE,
+    filing_data_ID bigint GENERATED ALWAYS AS IDENTITY UNIQUE,
+	filing_ID bigint REFERENCES html_extracts.sec_filing_id (filing_ID) ON DELETE CASCADE,
 	html_label TEXT NOT NULL,
     html_value TEXT NOT NULL,
 	/* period_begin DATE NOT NULL, */
